refactor: move static checks to their modules
